Move `type_caster<PyObject>` from test to new include (pybind11/type_caster_pyobject_ptr.h)
