+++ include/pybind11/cast.h
@@ -1041,10 +1041,11 @@ make_caster<T> load_type(const handle &handle) {
 PYBIND11_NAMESPACE_END(detail)
 
 // pytype -> C++ type
-template <
-    typename T,
-    detail::enable_if_t<!detail::is_pyobject<T>::value && !std::is_same<T, PyObject *>::value, int>
-    = 0>
+template <typename T,
+          detail::enable_if_t<!detail::is_pyobject<T>::value
+                                  && !detail::is_same_ignoring_cvref<T, PyObject *>::value,
+                              int>
+          = 0>
 T cast(const handle &handle) {
     using namespace detail;
     static_assert(!cast_is_temporary_value_reference<T>::value,
@@ -1059,7 +1060,7 @@ T cast(const handle &handle) {
 }
 
 template <typename T,
-          detail::enable_if_t<std::is_same<detail::remove_cvref_t<T>, PyObject *>::value, int> = 0>
+          detail::enable_if_t<detail::is_same_ignoring_cvref<T, PyObject *>::value, int> = 0>
 T cast(const handle &handle) {
     return handle.ptr();
 }
